docs/index.html

<!DOCTYPE html>
<html lang="en">
<head>
  <meta charset="UTF-8" />
  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
  <title>firefox-review — interactive demo</title>
  <link rel="stylesheet" href="../public/style.css" />
  <style>
    #demo-banner {
      background: #388bfd22;
      border-bottom: 1px solid #388bfd55;
      padding: 8px 24px;
      font-size: 13px;
      color: #79c0ff;
      display: flex;
      align-items: center;
      gap: 16px;
    }
    #demo-banner strong { color: #e6edf3; }
    #demo-banner a { color: #79c0ff; }
    #demo-banner .demo-note { color: #8b949e; }
  </style>
</head>
<body>

  <div id="demo-banner">
    <strong>Interactive demo</strong>
    <span class="demo-note">Fake patch data — all UI interactions work. Saves are no-ops.</span>
    <a href="https://github.com/alastor0325/firefox-review" style="margin-left:auto;">View on GitHub →</a>
  </div>

  <!-- ── Sticky top bar (header + prompt bar + patch tabs) ── -->
  <div id="top-bar">

    <!-- ── Update notification banner ── -->
    <div id="update-banner" style="display:none;">
      <span>Codebase updated — patches may have changed.</span>
      <button id="btn-reload-page">Reload</button>
    </div>

    <!-- ── Header ── -->
    <div id="header">
      <div id="header-left">
        <h1>firefox-review · <span id="bug-id-display"></span></h1>
        <div id="worktree-path"></div>
      </div>
      <div id="submit-area">
        <button id="btn-submit" disabled>Generate Review Prompt</button>
        <div style="display:flex;justify-content:space-between;align-items:center;gap:16px;">
          <span id="submit-warning">Review at least one patch before generating the prompt</span>
          <span id="autosave-status"></span>
        </div>
      </div>
    </div>

    <!-- ── Current prompt bar ── -->
    <div id="current-prompt-bar" style="display:none;">
      <span id="current-prompt-label">Review prompt ready</span>
      <button id="btn-copy-current-prompt">Copy current prompt</button>
    </div>

    <!-- ── Patch tabs ── -->
    <div id="patch-tabs-bar" style="display:none;">
      <div id="patch-tabs"></div>
    </div>

  </div>

  <!-- ── Loading / error ── -->
  <div id="loading">Loading diff…</div>
  <div id="error-msg" style="display:none;"></div>

  <!-- ── Files changed ── -->
  <div id="files-changed" style="display:none;"></div>

  <!-- ── Result overlay ── -->
  <div id="result-overlay">
    <div id="result-modal">
      <h3>Review saved!</h3>
      <p>Saved to: <code id="result-feedback-path"></code></p>
      <textarea id="result-prompt" readonly></textarea>
      <div style="display:flex;gap:8px;margin-top:12px;">
        <button id="btn-copy-prompt">Copy prompt</button>
        <button id="btn-close-modal">Close</button>
      </div>
    </div>
  </div>

  <script>
  // ── Demo mock data ──────────────────────────────────────────────────────────

  const MOCK_DIFF = {
    worktreeName: 'bug-1234567',
    worktreePath: '/Users/dev/firefox-bug-1234567',
    patches: [
      {
        hash: 'a1b2c3d4',
        message: 'Bug 1234567 - Part 1: Add waitingforkey IDL and event dispatch',
        body: 'Bug 1234567 - Part 1: Add waitingforkey IDL and event dispatch\n\nFire the waitingforkey event when encrypted media is encountered but no\nMediaKeys object is attached yet, per the EME spec §2.4.1.\n\nAlso expose onwaitingforkey as an event handler attribute in the IDL.',
        files: [
          {
            oldPath: 'dom/media/HTMLMediaElement.cpp',
            newPath: 'dom/media/HTMLMediaElement.cpp',
            binary: false,
            hunks: [
              {
                header: '@@ -108,6 +108,13 @@',
                oldStart: 108, oldCount: 6, newStart: 108, newCount: 13,
                lines: [
                  { type: 'context', content: 'void HTMLMediaElement::HandleKeyNeeded(', oldLineNum: 108, newLineNum: 108 },
                  { type: 'context', content: '    eme::MediaKeyNeededEventInit& aInit)', oldLineNum: 109, newLineNum: 109 },
                  { type: 'context', content: '{', oldLineNum: 110, newLineNum: 110 },
                  { type: 'added',   content: '  if (!mMediaKeys) {', oldLineNum: null, newLineNum: 111 },
                  { type: 'added',   content: '    // Fire waitingforkey per EME spec §2.4.1', oldLineNum: null, newLineNum: 112 },
                  { type: 'added',   content: '    DispatchAsyncEvent(u"waitingforkey"_ns);', oldLineNum: null, newLineNum: 113 },
                  { type: 'added',   content: '    mPendingKeyNeededEvents.AppendElement(aInit);', oldLineNum: null, newLineNum: 114 },
                  { type: 'added',   content: '    return;', oldLineNum: null, newLineNum: 115 },
                  { type: 'added',   content: '  }', oldLineNum: null, newLineNum: 116 },
                  { type: 'context', content: '  RefPtr<MediaKeySession> session;', oldLineNum: 111, newLineNum: 117 },
                  { type: 'context', content: '  mMediaKeys->GetKeyForRequest(aInit, getter_AddRefs(session));', oldLineNum: 112, newLineNum: 118 },
                  { type: 'context', content: '}', oldLineNum: 113, newLineNum: 119 },
                ],
              },
            ],
          },
          {
            oldPath: 'dom/media/HTMLMediaElement.webidl',
            newPath: 'dom/media/HTMLMediaElement.webidl',
            binary: false,
            hunks: [
              {
                header: '@@ -62,6 +62,7 @@',
                oldStart: 62, oldCount: 6, newStart: 62, newCount: 7,
                lines: [
                  { type: 'context', content: '  attribute EventHandler onencrypted;', oldLineNum: 62, newLineNum: 62 },
                  { type: 'added',   content: '  attribute EventHandler onwaitingforkey;', oldLineNum: null, newLineNum: 63 },
                  { type: 'context', content: '};', oldLineNum: 63, newLineNum: 64 },
                ],
              },
            ],
          },
        ],
      },
      {
        hash: 'e5f6a7b8',
        message: 'Bug 1234567 - Part 2: Handle ClearLead in ExternalEngineStateMachine',
        body: 'Bug 1234567 - Part 2: Handle ClearLead in ExternalEngineStateMachine\n\nTrack ClearLead state in ExternalEngineStateMachine so we know when a\nstream transitions from unencrypted to encrypted segments.\n\nSet mClearLeadState to ePossible when the media has encrypted tracks,\nand assert task-queue correctness in WaitingForKey.',
        files: [
          {
            oldPath: 'dom/media/ExternalEngineStateMachine.cpp',
            newPath: 'dom/media/ExternalEngineStateMachine.cpp',
            binary: false,
            hunks: [
              {
                header: '@@ -241,7 +241,15 @@',
                oldStart: 241, oldCount: 7, newStart: 241, newCount: 15,
                lines: [
                  { type: 'context', content: 'void ExternalEngineStateMachine::OnCDMProxyReady(', oldLineNum: 241, newLineNum: 241 },
                  { type: 'context', content: '    RefPtr<CDMProxy> aProxy)', oldLineNum: 242, newLineNum: 242 },
                  { type: 'context', content: '{', oldLineNum: 243, newLineNum: 243 },
                  { type: 'added',   content: '  MOZ_ASSERT(OnTaskQueue());', oldLineNum: null, newLineNum: 244 },
                  { type: 'added',   content: '  mClearLeadState = ClearLeadState::eNone;', oldLineNum: null, newLineNum: 245 },
                  { type: 'added',   content: '  if (mInfo.HasEncryptedTracks()) {', oldLineNum: null, newLineNum: 246 },
                  { type: 'added',   content: '    mClearLeadState = ClearLeadState::ePossible;', oldLineNum: null, newLineNum: 247 },
                  { type: 'added',   content: '  }', oldLineNum: null, newLineNum: 248 },
                  { type: 'context', content: '  mCDMProxy = std::move(aProxy);', oldLineNum: 244, newLineNum: 249 },
                  { type: 'context', content: '  mCDMProxy->SetEventTarget(mTaskQueue);', oldLineNum: 245, newLineNum: 250 },
                  { type: 'context', content: '}', oldLineNum: 246, newLineNum: 251 },
                ],
              },
              {
                header: '@@ -308,5 +316,8 @@',
                oldStart: 308, oldCount: 5, newStart: 316, newCount: 8,
                lines: [
                  { type: 'context', content: 'void ExternalEngineStateMachine::WaitingForKey()', oldLineNum: 308, newLineNum: 316 },
                  { type: 'context', content: '{', oldLineNum: 309, newLineNum: 317 },
                  { type: 'added',   content: '  MOZ_ASSERT(mCDMProxy);', oldLineNum: null, newLineNum: 318 },
                  { type: 'added',   content: '  MOZ_ASSERT(OnTaskQueue());', oldLineNum: null, newLineNum: 319 },
                  { type: 'context', content: '  SetState(DECODER_STATE_WAITING_FOR_KEY);', oldLineNum: 310, newLineNum: 320 },
                  { type: 'context', content: '}', oldLineNum: 311, newLineNum: 321 },
                ],
              },
            ],
          },
        ],
      },
      {
        hash: 'c9d0e1f2',
        message: 'Bug 1234567 - Part 3: Refresh CDM proxy trusted input on proxy set',
        body: 'Bug 1234567 - Part 3: Refresh CDM proxy trusted input on proxy set\n\nCall RefreshTrustedInput() on the CDM parent whenever a new CDM proxy is\nset, so the GPU process has up-to-date trusted input state.',
        files: [
          {
            oldPath: 'dom/media/platforms/wmf/MFCDMProxy.cpp',
            newPath: 'dom/media/platforms/wmf/MFCDMProxy.cpp',
            binary: false,
            hunks: [
              {
                header: '@@ -85,6 +85,14 @@',
                oldStart: 85, oldCount: 6, newStart: 85, newCount: 14,
                lines: [
                  { type: 'context', content: 'void MFCDMProxy::OnCDMProxySet(', oldLineNum: 85, newLineNum: 85 },
                  { type: 'context', content: '    RefPtr<MFCDMProxyCallback> aCallback)', oldLineNum: 86, newLineNum: 86 },
                  { type: 'context', content: '{', oldLineNum: 87, newLineNum: 87 },
                  { type: 'added',   content: '  MOZ_ASSERT(NS_IsMainThread());', oldLineNum: null, newLineNum: 88 },
                  { type: 'added',   content: '  if (const auto cdmParent = mCDMParent.get()) {', oldLineNum: null, newLineNum: 89 },
                  { type: 'added',   content: '    cdmParent->RefreshTrustedInput();', oldLineNum: null, newLineNum: 90 },
                  { type: 'added',   content: '  }', oldLineNum: null, newLineNum: 91 },
                  { type: 'removed', content: '  if (mCallback) {', oldLineNum: 88, newLineNum: null },
                  { type: 'removed', content: '    mCallback->OnCDMProxySet();', oldLineNum: 89, newLineNum: null },
                  { type: 'removed', content: '  }', oldLineNum: 90, newLineNum: null },
                  { type: 'added',   content: '  if (mCallback) { mCallback->OnCDMProxySet(); }', oldLineNum: null, newLineNum: 92 },
                  { type: 'context', content: '  mCallback = std::move(aCallback);', oldLineNum: 91, newLineNum: 93 },
                  { type: 'context', content: '}', oldLineNum: 92, newLineNum: 94 },
                ],
              },
            ],
          },
        ],
      },
      {
        hash: 'd3e4f5a6',
        message: 'Bug 1234567 - Part 4: Add GTest for WaitingForKey event dispatch',
        body: 'Bug 1234567 - Part 4: Add GTest for WaitingForKey event dispatch\n\nVerify that HandleKeyNeeded fires the waitingforkey event when no\nMediaKeys object is attached, and enqueues the init data.',
        files: [
          {
            oldPath: 'dom/media/gtest/TestEMEWaitingForKey.cpp',
            newPath: 'dom/media/gtest/TestEMEWaitingForKey.cpp',
            binary: false,
            hunks: [
              {
                header: '@@ -0,0 +1,32 @@',
                oldStart: 0, oldCount: 0, newStart: 1, newCount: 32,
                lines: [
                  { type: 'added', content: '/* -*- Mode: C++; tab-width: 8; indent-tabs-mode: nil -*- */', oldLineNum: null, newLineNum: 1 },
                  { type: 'added', content: '/* This Source Code Form is subject to the terms of the Mozilla Public', oldLineNum: null, newLineNum: 2 },
                  { type: 'added', content: ' * License, v. 2.0. If a copy of the MPL was not distributed with this', oldLineNum: null, newLineNum: 3 },
                  { type: 'added', content: ' * file, You can obtain one at http://mozilla.org/MPL/2.0/. */', oldLineNum: null, newLineNum: 4 },
                  { type: 'added', content: '#include "gtest/gtest.h"', oldLineNum: null, newLineNum: 5 },
                  { type: 'added', content: '#include "HTMLMediaElement.h"', oldLineNum: null, newLineNum: 6 },
                  { type: 'added', content: '', oldLineNum: null, newLineNum: 7 },
                  { type: 'added', content: 'TEST(EMEWaitingForKey, FiresWhenNoMediaKeys) {', oldLineNum: null, newLineNum: 8 },
                  { type: 'added', content: '  MockHTMLMediaElement el;', oldLineNum: null, newLineNum: 9 },
                  { type: 'added', content: '  eme::MediaKeyNeededEventInit init;', oldLineNum: null, newLineNum: 10 },
                  { type: 'added', content: '  EXPECT_FALSE(el.HasMediaKeys());', oldLineNum: null, newLineNum: 11 },
                  { type: 'added', content: '  el.HandleKeyNeeded(init);', oldLineNum: null, newLineNum: 12 },
                  { type: 'added', content: '  EXPECT_TRUE(el.WaitingForKeyEventFired());', oldLineNum: null, newLineNum: 13 },
                  { type: 'added', content: '  EXPECT_EQ(el.PendingKeyNeededEvents().Length(), 1u);', oldLineNum: null, newLineNum: 14 },
                  { type: 'added', content: '}', oldLineNum: null, newLineNum: 15 },
                ],
              },
            ],
          },
        ],
      },
      {
        hash: 'f7a8b9c0',
        message: 'Bug 1234567 - Part 5: Update mochitest for waitingforkey',
        body: 'Bug 1234567 - Part 5: Update mochitest for waitingforkey\n\nExtend the existing EME mochitest to cover the waitingforkey event\npath, ensuring the event fires before MediaKeys is set.',
        files: [
          {
            oldPath: 'dom/media/test/test_eme_waitingforkey.html',
            newPath: 'dom/media/test/test_eme_waitingforkey.html',
            binary: false,
            hunks: [
              {
                header: '@@ -44,6 +44,18 @@',
                oldStart: 44, oldCount: 6, newStart: 44, newCount: 18,
                lines: [
                  { type: 'context', content: '  // Verify encrypted event fires', oldLineNum: 44, newLineNum: 44 },
                  { type: 'context', content: '  await once(video, "encrypted");', oldLineNum: 45, newLineNum: 45 },
                  { type: 'context', content: '  ok(true, "encrypted event fired");', oldLineNum: 46, newLineNum: 46 },
                  { type: 'added',   content: '', oldLineNum: null, newLineNum: 47 },
                  { type: 'added',   content: '  // Verify waitingforkey fires before MediaKeys is attached', oldLineNum: null, newLineNum: 48 },
                  { type: 'added',   content: '  const wfkPromise = once(video, "waitingforkey");', oldLineNum: null, newLineNum: 49 },
                  { type: 'added',   content: '  video.src = EMEHelper.getEncryptedVideoSrc();', oldLineNum: null, newLineNum: 50 },
                  { type: 'added',   content: '  await wfkPromise;', oldLineNum: null, newLineNum: 51 },
                  { type: 'added',   content: '  ok(true, "waitingforkey fired before MediaKeys set");', oldLineNum: null, newLineNum: 52 },
                  { type: 'added',   content: '  await setMediaKeys(video);', oldLineNum: null, newLineNum: 53 },
                  { type: 'added',   content: '  ok(video.mediaKeys, "MediaKeys set after waitingforkey");', oldLineNum: null, newLineNum: 54 },
                  { type: 'context', content: '  SimpleTest.finish();', oldLineNum: 47, newLineNum: 55 },
                ],
              },
            ],
          },
        ],
      },
      {
        hash: 'b1c2d3e4',
        message: 'Bug 1234567 - Part 6: Update EME telemetry probes',
        body: 'Bug 1234567 - Part 6: Update EME telemetry probes\n\nAdd a new USE_COUNTER probe for waitingforkey usage so we can track\nhow often sites rely on the event in the wild.',
        files: [
          {
            oldPath: 'toolkit/components/telemetry/Histograms.json',
            newPath: 'toolkit/components/telemetry/Histograms.json',
            binary: false,
            hunks: [
              {
                header: '@@ -12450,6 +12450,14 @@',
                oldStart: 12450, oldCount: 6, newStart: 12450, newCount: 14,
                lines: [
                  { type: 'context', content: '  "EME_KEY_SESSION_CLOSE_EXCEPTION": {', oldLineNum: 12450, newLineNum: 12450 },
                  { type: 'context', content: '    "record_in_processes": ["main"],', oldLineNum: 12451, newLineNum: 12451 },
                  { type: 'context', content: '    "description": "Exceptions thrown by MediaKeySession.close()"', oldLineNum: 12452, newLineNum: 12452 },
                  { type: 'context', content: '  },', oldLineNum: 12453, newLineNum: 12453 },
                  { type: 'added',   content: '  "EME_WAITING_FOR_KEY_EVENT_FIRED": {', oldLineNum: null, newLineNum: 12454 },
                  { type: 'added',   content: '    "record_in_processes": ["main"],', oldLineNum: null, newLineNum: 12455 },
                  { type: 'added',   content: '    "kind": "boolean",', oldLineNum: null, newLineNum: 12456 },
                  { type: 'added',   content: '    "description": "Whether the waitingforkey event was fired during EME playback"', oldLineNum: null, newLineNum: 12457 },
                  { type: 'added',   content: '  },', oldLineNum: null, newLineNum: 12458 },
                  { type: 'context', content: '  "EME_GENERATE_REQUEST_FAILED": {', oldLineNum: 12454, newLineNum: 12459 },
                ],
              },
            ],
          },
        ],
      },
    ],
  };

  // Rev 2 for Part 2 (added MOZ_ASSERT + basic ClearLead tracking, before the full implementation)
  const MOCK_OLD_DIFF_PART2_REV2 = {
    hash: 'old_e5f6',
    files: [
      {
        oldPath: 'dom/media/ExternalEngineStateMachine.cpp',
        newPath: 'dom/media/ExternalEngineStateMachine.cpp',
        binary: false,
        hunks: [
          {
            header: '@@ -241,5 +241,8 @@',
            oldStart: 241, oldCount: 5, newStart: 241, newCount: 8,
            lines: [
              { type: 'context', content: 'void ExternalEngineStateMachine::OnCDMProxyReady(', oldLineNum: 241, newLineNum: 241 },
              { type: 'context', content: '    RefPtr<CDMProxy> aProxy)', oldLineNum: 242, newLineNum: 242 },
              { type: 'context', content: '{', oldLineNum: 243, newLineNum: 243 },
              { type: 'added',   content: '  MOZ_ASSERT(OnTaskQueue());', oldLineNum: null, newLineNum: 244 },
              { type: 'context', content: '  mCDMProxy = std::move(aProxy);', oldLineNum: 244, newLineNum: 245 },
              { type: 'context', content: '  mCDMProxy->SetEventTarget(mTaskQueue);', oldLineNum: 245, newLineNum: 246 },
              { type: 'context', content: '}', oldLineNum: 246, newLineNum: 247 },
            ],
          },
        ],
      },
    ],
  };

  // Rev 1 for Part 2 (original — no MOZ_ASSERT, no ClearLead tracking at all)
  const MOCK_OLD_DIFF_PART2_REV1 = {
    hash: 'rev1_e5f6',
    files: [
      {
        oldPath: 'dom/media/ExternalEngineStateMachine.cpp',
        newPath: 'dom/media/ExternalEngineStateMachine.cpp',
        binary: false,
        hunks: [
          {
            header: '@@ -241,4 +241,6 @@',
            oldStart: 241, oldCount: 4, newStart: 241, newCount: 6,
            lines: [
              { type: 'context', content: 'void ExternalEngineStateMachine::OnCDMProxyReady(', oldLineNum: 241, newLineNum: 241 },
              { type: 'context', content: '    RefPtr<CDMProxy> aProxy)', oldLineNum: 242, newLineNum: 242 },
              { type: 'context', content: '{', oldLineNum: 243, newLineNum: 243 },
              { type: 'added',   content: '  mCDMProxy = std::move(aProxy);', oldLineNum: null, newLineNum: 244 },
              { type: 'added',   content: '  mCDMProxy->SetEventTarget(mTaskQueue);', oldLineNum: null, newLineNum: 245 },
              { type: 'context', content: '}', oldLineNum: 244, newLineNum: 246 },
            ],
          },
        ],
      },
    ],
  };

  const MOCK_STATE = {
    comments: {
      'e5f6a7b8': {
        'dom/media/ExternalEngineStateMachine.cpp': {
          'n245': {
            patchHash: 'e5f6a7b8',
            file: 'dom/media/ExternalEngineStateMachine.cpp',
            line: 245,
            lineContent: '  mClearLeadState = ClearLeadState::eNone;',
            text: 'Should this reset to eNone or ePossible here? The check on line 246 overrides this anyway when HasEncryptedTracks().',
          },
          'n318': {
            patchHash: 'e5f6a7b8',
            file: 'dom/media/ExternalEngineStateMachine.cpp',
            line: 318,
            lineContent: '  MOZ_ASSERT(mCDMProxy);',
            text: 'Add a failure message: MOZ_ASSERT(mCDMProxy, "CDM proxy must be set before waiting for key")',
          },
        },
      },
      'c9d0e1f2': {
        '__commit__': {
          'msg': {
            patchHash: 'c9d0e1f2',
            file: '__commit__',
            line: 0,
            lineContent: 'Bug 1234567 - Part 3: Refresh CDM proxy trusted input on proxy set',
            text: 'Commit message should describe WHY we need to refresh trusted input, not just what we do. E.g. "…so the GPU process reflects the latest CDM state."',
          },
        },
      },
      'b1c2d3e4': {
        'toolkit/components/telemetry/Histograms.json': {
          'n12457': {
            patchHash: 'b1c2d3e4',
            file: 'toolkit/components/telemetry/Histograms.json',
            line: 12457,
            lineContent: '    "description": "Whether the waitingforkey event was fired during EME playback"',
            text: 'Telemetry probes for EME events need a data review. Please file a separate bug and link it here.',
          },
        },
      },
    },
    generalComments: {
      'c9d0e1f2': 'The removal and re-addition of the callback check is unnecessary churn. Please only add the RefreshTrustedInput() call without restructuring the existing code.',
    },
    approved: ['a1b2c3d4', 'd3e4f5a6', 'f7a8b9c0'],
    denied: ['e5f6a7b8', 'c9d0e1f2', 'b1c2d3e4'],
    // Revision history — Part 2 has been amended twice
    revisions: [
      {
        savedAt: '2025-01-12T14:30:00Z',
        patches: [
          { hash: 'a1b2c3d4', message: 'Bug 1234567 - Part 1: Add waitingforkey IDL and event dispatch' },
          { hash: 'rev1_e5f6', message: 'Bug 1234567 - Part 2: Handle ClearLead in ExternalEngineStateMachine' },
          { hash: 'c9d0e1f2', message: 'Bug 1234567 - Part 3: Refresh CDM proxy trusted input on proxy set' },
          { hash: 'd3e4f5a6', message: 'Bug 1234567 - Part 4: Add GTest for WaitingForKey event dispatch' },
          { hash: 'f7a8b9c0', message: 'Bug 1234567 - Part 5: Update mochitest for waitingforkey' },
          { hash: 'b1c2d3e4', message: 'Bug 1234567 - Part 6: Update EME telemetry probes' },
        ],
      },
      {
        savedAt: '2025-01-14T09:15:00Z',
        patches: [
          { hash: 'a1b2c3d4', message: 'Bug 1234567 - Part 1: Add waitingforkey IDL and event dispatch' },
          { hash: 'old_e5f6', message: 'Bug 1234567 - Part 2: Handle ClearLead in ExternalEngineStateMachine' },
          { hash: 'c9d0e1f2', message: 'Bug 1234567 - Part 3: Refresh CDM proxy trusted input on proxy set' },
          { hash: 'd3e4f5a6', message: 'Bug 1234567 - Part 4: Add GTest for WaitingForKey event dispatch' },
          { hash: 'f7a8b9c0', message: 'Bug 1234567 - Part 5: Update mochitest for waitingforkey' },
          { hash: 'b1c2d3e4', message: 'Bug 1234567 - Part 6: Update EME telemetry probes' },
        ],
      },
      {
        savedAt: '2025-01-15T16:42:00Z',
        patches: [
          { hash: 'a1b2c3d4', message: 'Bug 1234567 - Part 1: Add waitingforkey IDL and event dispatch' },
          { hash: 'e5f6a7b8', message: 'Bug 1234567 - Part 2: Handle ClearLead in ExternalEngineStateMachine' },
          { hash: 'c9d0e1f2', message: 'Bug 1234567 - Part 3: Refresh CDM proxy trusted input on proxy set' },
          { hash: 'd3e4f5a6', message: 'Bug 1234567 - Part 4: Add GTest for WaitingForKey event dispatch' },
          { hash: 'f7a8b9c0', message: 'Bug 1234567 - Part 5: Update mochitest for waitingforkey' },
          { hash: 'b1c2d3e4', message: 'Bug 1234567 - Part 6: Update EME telemetry probes' },
        ],
      },
    ],
    prompt: 'You are being asked to revise your implementation in worktree firefox-bug-1234567.\n\n## Full patch series:\n- a1b2c3d4 Bug 1234567 - Part 1: Add waitingforkey IDL and event dispatch  [APPROVED — no issues]\n- e5f6a7b8 Bug 1234567 - Part 2: Handle ClearLead in ExternalEngineStateMachine  [DENIED — requires significant changes]\n- c9d0e1f2 Bug 1234567 - Part 3: Refresh CDM proxy trusted input on proxy set  [DENIED — requires significant changes]\n- d3e4f5a6 Bug 1234567 - Part 4: Add GTest for WaitingForKey event dispatch  [APPROVED — no issues]\n- f7a8b9c0 Bug 1234567 - Part 5: Update mochitest for waitingforkey  [APPROVED — no issues]\n- b1c2d3e4 Bug 1234567 - Part 6: Update EME telemetry probes  [DENIED — requires significant changes]',
  };

  // ── Mock fetch ────────────────────────────────────────────────────────────

  window.fetch = function(url, opts) {
    // GET /api/diff
    if (url === '/api/diff') {
      return Promise.resolve({
        ok: true,
        json: () => Promise.resolve(MOCK_DIFF),
      });
    }
    // GET /api/state
    if (url === '/api/state' && (!opts || opts.method !== 'POST')) {
      return Promise.resolve({
        ok: true,
        json: () => Promise.resolve(MOCK_STATE),
      });
    }
    // POST /api/state (auto-save — no-op in demo)
    if (url === '/api/state' && opts && opts.method === 'POST') {
      return Promise.resolve({
        ok: true,
        json: () => Promise.resolve({ ok: true }),
      });
    }
    // POST /api/submit
    if (url === '/api/submit') {
      return Promise.resolve({
        ok: true,
        json: () => Promise.resolve({
          ok: true,
          feedbackPath: '/Users/dev/firefox-bug-1234567/REVIEW_FEEDBACK_bug-1234567.md',
          prompt: 'Please read the review feedback at /Users/dev/firefox-bug-1234567/REVIEW_FEEDBACK_bug-1234567.md and apply all the changes described there to your implementation in worktree firefox-bug-1234567.',
        }),
      });
    }
    // GET /api/revdiff?from=X&to=Y (compare two revisions)
    if (url.startsWith('/api/revdiff?')) {
      const params = new URLSearchParams(url.split('?')[1]);
      const from = params.get('from');
      const to   = params.get('to');
      // Mock: Rev1→Rev2 shows the MOZ_ASSERT addition; anything else is empty
      const files = (from === 'rev1_e5f6' && to === 'old_e5f6') ? [
        {
          oldPath: 'dom/media/ExternalEngineStateMachine.cpp',
          newPath: 'dom/media/ExternalEngineStateMachine.cpp',
          binary: false,
          hunks: [
            {
              header: '@@ -241,6 +241,7 @@',
              oldStart: 241, oldCount: 6, newStart: 241, newCount: 7,
              lines: [
                { type: 'context', content: 'void ExternalEngineStateMachine::OnCDMProxyReady(', oldLineNum: 241, newLineNum: 241 },
                { type: 'context', content: '    RefPtr<CDMProxy> aProxy)', oldLineNum: 242, newLineNum: 242 },
                { type: 'context', content: '{', oldLineNum: 243, newLineNum: 243 },
                { type: 'added',   content: '  MOZ_ASSERT(OnTaskQueue());', oldLineNum: null, newLineNum: 244 },
                { type: 'context', content: '  mCDMProxy = std::move(aProxy);', oldLineNum: 244, newLineNum: 245 },
                { type: 'context', content: '  mCDMProxy->SetEventTarget(mTaskQueue);', oldLineNum: 245, newLineNum: 246 },
                { type: 'context', content: '}', oldLineNum: 246, newLineNum: 247 },
              ],
            },
          ],
        },
      ] : (from === 'old_e5f6' && to === 'e5f6a7b8') ? [
        {
          oldPath: 'dom/media/ExternalEngineStateMachine.cpp',
          newPath: 'dom/media/ExternalEngineStateMachine.cpp',
          binary: false,
          hunks: [
            {
              header: '@@ -241,7 +241,15 @@',
              oldStart: 241, oldCount: 7, newStart: 241, newCount: 15,
              lines: [
                { type: 'context', content: 'void ExternalEngineStateMachine::OnCDMProxyReady(', oldLineNum: 241, newLineNum: 241 },
                { type: 'context', content: '    RefPtr<CDMProxy> aProxy)', oldLineNum: 242, newLineNum: 242 },
                { type: 'context', content: '{', oldLineNum: 243, newLineNum: 243 },
                { type: 'context', content: '  MOZ_ASSERT(OnTaskQueue());', oldLineNum: 244, newLineNum: 244 },
                { type: 'added',   content: '  mClearLeadState = ClearLeadState::eNone;', oldLineNum: null, newLineNum: 245 },
                { type: 'added',   content: '  if (mInfo.HasEncryptedTracks()) {', oldLineNum: null, newLineNum: 246 },
                { type: 'added',   content: '    mClearLeadState = ClearLeadState::ePossible;', oldLineNum: null, newLineNum: 247 },
                { type: 'added',   content: '  }', oldLineNum: null, newLineNum: 248 },
                { type: 'context', content: '  mCDMProxy = std::move(aProxy);', oldLineNum: 245, newLineNum: 249 },
                { type: 'context', content: '  mCDMProxy->SetEventTarget(mTaskQueue);', oldLineNum: 246, newLineNum: 250 },
                { type: 'context', content: '}', oldLineNum: 247, newLineNum: 251 },
              ],
            },
          ],
        },
      ] : (from === 'rev1_e5f6' && to === 'e5f6a7b8') ? [
        {
          oldPath: 'dom/media/ExternalEngineStateMachine.cpp',
          newPath: 'dom/media/ExternalEngineStateMachine.cpp',
          binary: false,
          hunks: [
            {
              header: '@@ -241,6 +241,15 @@',
              oldStart: 241, oldCount: 6, newStart: 241, newCount: 15,
              lines: [
                { type: 'context', content: 'void ExternalEngineStateMachine::OnCDMProxyReady(', oldLineNum: 241, newLineNum: 241 },
                { type: 'context', content: '    RefPtr<CDMProxy> aProxy)', oldLineNum: 242, newLineNum: 242 },
                { type: 'context', content: '{', oldLineNum: 243, newLineNum: 243 },
                { type: 'added',   content: '  MOZ_ASSERT(OnTaskQueue());', oldLineNum: null, newLineNum: 244 },
                { type: 'added',   content: '  mClearLeadState = ClearLeadState::eNone;', oldLineNum: null, newLineNum: 245 },
                { type: 'added',   content: '  if (mInfo.HasEncryptedTracks()) {', oldLineNum: null, newLineNum: 246 },
                { type: 'added',   content: '    mClearLeadState = ClearLeadState::ePossible;', oldLineNum: null, newLineNum: 247 },
                { type: 'added',   content: '  }', oldLineNum: null, newLineNum: 248 },
                { type: 'context', content: '  mCDMProxy = std::move(aProxy);', oldLineNum: 244, newLineNum: 249 },
                { type: 'context', content: '  mCDMProxy->SetEventTarget(mTaskQueue);', oldLineNum: 245, newLineNum: 250 },
                { type: 'context', content: '}', oldLineNum: 246, newLineNum: 251 },
              ],
            },
          ],
        },
      ] : [];
      return Promise.resolve({ ok: true, json: () => Promise.resolve({ from, to, files }) });
    }
    // GET /api/headhash (update detection — always returns same hash in demo)
    if (url === '/api/headhash') {
      return Promise.resolve({
        ok: true,
        json: () => Promise.resolve({ hash: 'a1b2c3d4' }),
      });
    }
    // GET /api/patchdiff/:hash (historical revision diffs)
    if (url.startsWith('/api/patchdiff/')) {
      const hash = url.split('/api/patchdiff/')[1];
      const diffMap = {
        'old_e5f6':  MOCK_OLD_DIFF_PART2_REV2,
        'rev1_e5f6': MOCK_OLD_DIFF_PART2_REV1,
      };
      const diff = diffMap[hash] || { hash, files: [], error: `No mock diff for ${hash}` };
      return Promise.resolve({ ok: true, json: () => Promise.resolve(diff) });
    }
    // Fallback — should not happen in demo
    return Promise.reject(new Error(`Unmocked fetch: ${url}`));
  };
  </script>

  <script src="../public/app.js"></script>
</body>
</html>
